LayoutTests/imported/w3c/web-platform-tests/shadow-dom/event-composed-path-after-dom-mutation-expected.txt


PASS Event.composedPath() should return the same result even if DOM is mutated (1/2)
PASS Event.composedPath() should return the same result even if DOM is mutated (2/2)
